refactor(account): relocate getAccountLabelById function to utils for better organization
